♻️ refactor: refactor modal
