+++ src/app/(main)/settings/llm/components/ProviderModelList/ModelConfigModal/Form.tsx
@@ -1,7 +1,6 @@
-import { Modal } from '@lobehub/ui';
-import { Button, Checkbox, Form, Input } from 'antd';
+import { Checkbox, Form, FormInstance, Input } from 'antd';
 import isEqual from 'fast-deep-equal';
-import { memo } from 'react';
+import { memo, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { useUserStore } from '@/store/user';
@@ -9,63 +8,32 @@ import { modelConfigSelectors } from '@/store/user/selectors';
 
 import MaxTokenSlider from './MaxTokenSlider';
 
-interface ModelConfigModalProps {
-  provider?: string;
+interface ModelConfigFormProps {
+  onFormInstanceReady: (instance: FormInstance) => void;
   showAzureDeployName?: boolean;
 }
-const ModelConfigModal = memo<ModelConfigModalProps>(({ showAzureDeployName, provider }) => {
-  const [formInstance] = Form.useForm();
-  const { t } = useTranslation('setting');
-  const { t: tc } = useTranslation('common');
 
-  const [open, id, editingProvider, dispatchCustomModelCards, toggleEditingCustomModelCard] =
-    useUserStore((s) => [
-      !!s.editingCustomCardModel && provider === s.editingCustomCardModel?.provider,
+const ModelConfigForm = memo<ModelConfigFormProps>(
+  ({ showAzureDeployName, onFormInstanceReady }) => {
+    const { t } = useTranslation('setting');
+
+    const [id, editingProvider] = useUserStore((s) => [
       s.editingCustomCardModel?.id,
       s.editingCustomCardModel?.provider,
-      s.dispatchCustomModelCards,
-      s.toggleEditingCustomModelCard,
     ]);
 
-  const modelCard = useUserStore(
-    modelConfigSelectors.getCustomModelCard({ id, provider: editingProvider }),
-    isEqual,
-  );
+    const modelCard = useUserStore(
+      modelConfigSelectors.getCustomModelCard({ id, provider: editingProvider }),
+      isEqual,
+    );
 
-  const closeModal = () => {
-    toggleEditingCustomModelCard(undefined);
-  };
+    const [formInstance] = Form.useForm();
 
-  return (
-    <Modal
-      destroyOnClose
-      footer={[
-        <Button key="cancel" onClick={closeModal}>
-          {tc('cancel')}
-        </Button>,
+    useEffect(() => {
+      onFormInstanceReady(formInstance);
+    }, []);
 
-        <Button
-          key="ok"
-          onClick={() => {
-            if (!editingProvider || !id) return;
-            const data = formInstance.getFieldsValue();
-
-            dispatchCustomModelCards(editingProvider as any, { id, type: 'update', value: data });
-
-            closeModal();
-          }}
-          style={{ marginInlineStart: '16px' }}
-          type="primary"
-        >
-          {tc('ok')}
-        </Button>,
-      ]}
-      maskClosable
-      onCancel={closeModal}
-      open={open}
-      title={t('llm.customModelCards.modelConfig.modalTitle')}
-      zIndex={1051} // Select is 1050
-    >
+    return (
       <div
         onClick={(e) => {
           e.stopPropagation();
@@ -79,7 +47,6 @@ const ModelConfigModal = memo<ModelConfigModalProps>(({ showAzureDeployName, pro
           form={formInstance}
           initialValues={modelCard}
           labelCol={{ span: 4 }}
-          preserve={false}
           style={{ marginTop: 16 }}
           wrapperCol={{ offset: 1, span: 18 }}
         >
@@ -136,7 +103,7 @@ const ModelConfigModal = memo<ModelConfigModalProps>(({ showAzureDeployName, pro
           </Form.Item>
         </Form>
       </div>
-    </Modal>
-  );
-});
-export default ModelConfigModal;
+    );
+  },
+);
+export default ModelConfigForm;
